requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
bcrypt==4.0.1
boto3==1.28.20
botocore==1.31.20
CacheControl==0.13.1
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.3
Django==4.2.5
django-crontab==0.7.1
django-environ==0.11.1
django-redis==5.3.0
djangorestframework==3.14.0
fcm-django==2.0.0
firebase-admin==6.2.0
google-api-core==2.12.0.dev1
google-api-python-client==2.97.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.3
google-cloud-firestore==2.11.1
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.60.0
grpcio==1.58.0rc1
grpcio-status==1.58.0rc1
gunicorn==21.2.0
httplib2==0.22.0
idna==3.4
jmespath==1.0.1
msgpack==1.0.5
mysqlclient==2.2.0
packaging==23.1
proto-plus==1.22.3
protobuf==4.24.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3
redis==5.0.0
requests==2.31.0
rsa==4.9
s3transfer==0.6.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==1.26.16
uWSGI==2.0.22
whitenoise==6.5.0
